requirements.txt
pycardano
requests
python-dotenv
gspreads